:building_construction: Fix: File Structer
파일 이름 변경 및 통합
