+++ src/lib/api/accont/index.api.ts
@@ -1,0 +1,15 @@
+import { CustomAxios } from "src/lib/Axios"
+
+export const handleOpenAccount = async (req: object) => {
+  const url: string = '/account';
+  const { data } = await CustomAxios.post(url, req);
+
+  return data;
+}
+
+export const handleGetMyAccount = async () => {
+  const url: string = '/account/find/my';
+  const { data } = await CustomAxios.get(url);
+
+  return data
+}
