--- src/lib/api/auth/login.api.ts
@@ -1,8 +1,0 @@
-import { CustomAxios } from "src/lib/Axios";
-
-export const handleLogin = async (req: object) => {
-  const url: string = '/auth/login';
-  const { data } = await CustomAxios.post(url, req);
-
-  return data;
-}
